+++ src/primitives/sphere.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   sphere.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bwerner <bwerner@student.42heilbronn.de>   +#+  +:+       +#+        */
+/*   By: nmihaile <nmihaile@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/03 21:30:03 by nmihaile          #+#    #+#             */
-/*   Updated: 2024/08/10 23:05:15 by bwerner          ###   ########.fr       */
+/*   Updated: 2024/08/11 10:43:22 by nmihaile         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ static float	get_discriminant(t_ray ray, t_object *object, float *t0, float *t1)
 	discriminant = B * B - 4 * A * (vec3_dot(AP, AP) - object->radius * object->radius);
 	if (discriminant < 0)
 		return (discriminant);
-	sqrt_discriminant = sqrt(discriminant);
+	sqrt_discriminant = sqrtf(discriminant);
 	*t0 = (-B + sqrt_discriminant) / (2 * A);
 	*t1 = (-B - sqrt_discriminant) / (2 * A);
 	return (discriminant);
