+++ src/utils/vec3_rotation.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vec3_rotation.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bwerner <bwerner@student.42heilbronn.de>   +#+  +:+       +#+        */
+/*   By: nmihaile <nmihaile@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/02 14:47:52 by nmihaile          #+#    #+#             */
-/*   Updated: 2024/08/10 20:14:58 by bwerner          ###   ########.fr       */
+/*   Updated: 2024/08/11 10:44:56 by nmihaile         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ t_vec3	vec3_rotate_x(t_vec3 v, float rad)
 	float	s;
 	float	c;
 
-	s = sin(rad);
-	c = cos(rad);
+	s = sinf(rad);
+	c = cosf(rad);
 	return ((t_vec3){v.x, v.y * c - v.z * s, v.y * s + v.z * c});
 }
 
@@ -27,8 +27,8 @@ t_vec3	vec3_rotate_y(t_vec3 v, float rad)
 	float	s;
 	float	c;
 
-	s = sin(rad);
-	c = cos(rad);
+	s = sinf(rad);
+	c = cosf(rad);
 	return ((t_vec3){v.x * c + v.z * s, v.y, -v.x * s + v.z * c});
 }
 
@@ -37,7 +37,7 @@ t_vec3	vec3_rotate_z(t_vec3 v, float rad)
 	float	s;
 	float	c;
 
-	s = sin(rad);
-	c = cos(rad);
+	s = sinf(rad);
+	c = cosf(rad);
 	return ((t_vec3){v.x * c - v.y * s, v.x * s + v.y * c, v.z});
 }
